[FIX][ADD] [#155] 기능을 분할하기 위해 파일 분할 및 간단한 Footer Btn 추가
- Footer Btn 추가
- 파일을 나누어서 책임을 분할
- parseJwt 를 만들어 FE에서 사용자 분류
